backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
supabase>=2.24.0
openai==1.12.0
pypdf2==3.0.1
pdfplumber==0.10.3
numpy==1.26.3
python-multipart==0.0.6
pydantic==2.5.3
httpx>=0.24.0,<0.28
websockets>=15.0

